evm ee stf primitives
